requirements.txt
Pillow==10.1.0
openai==1.3.4
